add converter interactor
